# school-results-db/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Results & Examination API",
    "description": "Complete backend API service for Examination, Marks Management, Lifecycles, and Frozen Report Cards",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/marks/sheet": {
      "get": {
        "summary": "Get marks entry sheet for a class, subject, and section",
        "tags": [
          "Marks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "examClassId",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "classSubjectId",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "sectionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/marks/bulk": {
      "put": {
        "summary": "Bulk save marks for components and enrollments",
        "tags": [
          "Marks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "marks"
                ],
                "properties": {
                  "marks": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "componentId",
                        "enrollmentId",
                        "status"
                      ],
                      "properties": {
                        "componentId": {
                          "type": "number"
                        },
                        "enrollmentId": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "PRESENT",
                            "ABSENT",
                            "EXEMPT",
                            "MEDICAL_LEAVE",
                            "WITHHELD",
                            "NOT_APPLICABLE"
                          ]
                        },
                        "marksObtained": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "graceMarks": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "graceReason": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "remarks": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/marks/missing": {
      "get": {
        "summary": "Get all missing marks entries for an exam class",
        "tags": [
          "Marks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "examClassId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/report-cards": {
      "get": {
        "summary": "Get latest report card snapshot for a student",
        "tags": [
          "Report Cards"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "admissionNo",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "yearLabel",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "examName",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/report-cards/transcript": {
      "get": {
        "summary": "Get multi-year academic transcript from snapshots",
        "tags": [
          "Report Cards"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "admissionNo",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/report-cards/{snapshotId}/verify": {
      "get": {
        "summary": "Verify report card snapshot authenticity and integrity against SHA-256 hash",
        "tags": [
          "Report Cards"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "snapshotId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}